+++ Include/internal/pycore_atomic_funcs.h
@@ -49,9 +49,9 @@ static inline Py_ssize_t _Py_atomic_size_get(Py_ssize_t *var)
         old = *volatile_var;
     } while(_InterlockedCompareExchange64(volatile_var, old, old) != old);
 #else
-    Py_BUILD_ASSERT(sizeof(long) == sizeof(*var));
-    volatile long *volatile_var = (volatile long *)var;
-    long old;
+    Py_BUILD_ASSERT(sizeof(int) == sizeof(*var));
+    volatile int *volatile_var = (volatile int *)var;
+    int old;
     do {
         old = *volatile_var;
     } while(_InterlockedCompareExchange(volatile_var, old, old) != old);
@@ -66,8 +66,8 @@ static inline void _Py_atomic_size_set(Py_ssize_t *var, Py_ssize_t value)
     volatile __int64 *volatile_var = (volatile __int64 *)var;
     _InterlockedExchange64(volatile_var, value);
 #else
-    Py_BUILD_ASSERT(sizeof(long) == sizeof(*var));
-    volatile long *volatile_var = (volatile long *)var;
+    Py_BUILD_ASSERT(sizeof(int) == sizeof(*var));
+    volatile int *volatile_var = (volatile int *)var;
     _InterlockedExchange(volatile_var, value);
 #endif
 }
