+++ Include/internal/pycore_runtime.h
@@ -98,7 +98,7 @@ typedef struct pyruntimestate {
         struct _xidregitem *head;
     } xidregistry;
 
-    unsigned long main_thread;
+    unsigned int main_thread;
 
 #define NEXITFUNCS 32
     void (*exitfuncs[NEXITFUNCS])(void);
