+++ Include/pyexpat.h
@@ -49,7 +49,7 @@ struct PyExpat_CAPI
     int (*DefaultUnknownEncodingHandler)(
         void *encodingHandlerData, const XML_Char *name, XML_Encoding *info);
     /* might be none for expat < 2.1.0 */
-    int (*SetHashSalt)(XML_Parser parser, unsigned long hash_salt);
+    int (*SetHashSalt)(XML_Parser parser, unsigned int hash_salt);
     /* always add new stuff to the end! */
 };
